app/requirements.txt
streamlit==1.32.0
duckdb==0.10.0
pandas==2.2.0
plotly==5.20.0
anthropic==0.18.1
python-dotenv==1.0.0
loguru==0.7.2
